Fixed drawFrame runnable posting stupidness
Telling the engine to iterate is much more sane now. I also realized
that the engine is destroyed and recreated when "set wallpaper" is
selected after the preview. This was causing the engine to get stuck in
a uninitialized state as it was waiting to be told that the cat
transport had completed. Which the service only does when itself is
created. I've got a fix in that works, but I'm not super happy with it.
However, it does allow me to finally polish the actual visual of the
wallpaper.